requirements.txt
discord.py==2.3.2
gTTS==2.5.1
requests==2.32.3
